Added --append option
By default graffiti-monkey overwrites existing tags when propagating.
Setting this option makes it append the new tags to the list of
existing tags. Note that AWS enforces a maximum of ten tags per
resource and if this limit is exceeded only the first ten tags (with
existing tags taking priority) are set. The tag snapshot function is
wrapped in a try.. except block to handle this.